# Improved performance of filter performance via Simd selection [3x]