--- a/src/types/bit_chunk.rs
+++ b/src/types/bit_chunk.rs
@@ -99,6 +99,10 @@ impl<T: BitChunk> Iterator for BitChunkIter<T> {
     }
 }
 
+// # Safety
+// a mathematical invariant of this iterator
+unsafe impl<T: BitChunk> crate::trusted_len::TrustedLen for BitChunkIter<T> {}
+
 /// An [`Iterator<Item=usize>`] over a [`BitChunk`].
 /// This iterator returns the postion of bit set.
 /// Refer: https://lemire.me/blog/2018/03/08/iterating-over-set-bits-quickly-simd-edition/
@@ -147,6 +151,10 @@ impl<T: BitChunk> Iterator for BitChunkOnes<T> {
     }
 }
 
+// # Safety
+// a mathematical invariant of this iterator
+unsafe impl<T: BitChunk> crate::trusted_len::TrustedLen for BitChunkOnes<T> {}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -165,6 +173,7 @@ mod tests {
         let a = [0b00000001, 0b00010000]; // 0th and 13th entry
         let a = u16::from_ne_bytes(a);
         let mut iter = BitChunkOnes::new(a);
+        assert_eq!(iter.size_hint(), (2, Some(2)));
         assert_eq!(iter.next(), Some(0));
         assert_eq!(iter.next(), Some(12));
     }