Benchmarks:

Before:

```rust
filter f32              time:   [269.62 us 269.88 us 270.12 us]                       
filter f32 high selectivity                                                                            
                        time:   [243.83 us 244.41 us 244.95 us]
filter f32 nonull       time:   [108.89 us 108.95 us 109.03 us]                              
filter f32 nonull high selectivity                                                                             
                        time:   [18.886 us 18.909 us 18.933 us]
 
```

Now:

``` 
filter f32            time:   [180.22 us 180.34 us 180.45 us]                       
filter f32 high selectivity                                                                            
                        time:   [94.458 us 94.495 us 94.540 us]
			
filter f32 nonull       time:   [58.747 us 58.775 us 58.809 us]                              
filter f32 nonull high selectivity                                                                             
                        time:   [10.456 us 10.477 us 10.494 us]
```